Added demo and renamed the header.
